# service-core/docs/swagger.yaml
definitions:
  controllers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.Campaign:
    description: Represents an SMS campaign associated with a user
    properties:
      campaign_name:
        description: Campaign_Name is the name of the campaign
        type: string
      campaign_recipients:
        description: CampaignRecipients represents the recipients associated with
          this campaign
        items:
          $ref: '#/definitions/models.CampaignRecipient'
        type: array
      campaign_workflow_processings:
        description: CampaignWorkflowProcessings represents the workflows associated
          with this campaign
        items:
          $ref: '#/definitions/models.CampaignWorkflowProcessing'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      end_date:
        description: End_Date is the end date of the campaign
        type: string
      id:
        type: string
      message_body:
        description: Message_Body is the body of the SMS message for the campaign
        type: string
      start_date:
        description: Start_Date is the start date of the campaign
        type: string
      status:
        description: Status indicates the status of the campaign (e.g., Active, Inactive)
        type: string
      updated_at:
        type: string
      user_id:
        description: UserID is the ID of the user who created the campaign
        type: integer
    type: object
  models.CampaignRecipient:
    description: Represents a recipient of a campaign and their delivery status
    properties:
      campaign:
        allOf:
        - $ref: '#/definitions/models.Campaign'
        description: Campaign represents the campaign associated with this recipient
      campaign_id:
        description: CampaignID is the ID of the campaign (foreign key)
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      recipient:
        description: RecipientID is the ID of the recipient (foreign key)
        type: integer
      status:
        description: Status indicates the delivery status of the campaign to the recipient
        type: string
      updated_at:
        type: string
    type: object
  models.CampaignWorkflow:
    description: Represents a workflow for a campaign
    properties:
      campaign_workflow_processings:
        description: CampaignWorkflowProcessings represents the campaigns associated
          with this workflow
        items:
          $ref: '#/definitions/models.CampaignWorkflowProcessing'
        type: array
      campaign_workflow_users:
        description: CampaignWorkflowUsers represents the users associated with this
          workflow
        items:
          $ref: '#/definitions/models.CampaignWorkflowUser'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        description: Name is the name of the workflow
        type: string
      updated_at:
        type: string
    type: object
  models.CampaignWorkflowProcessing:
    description: Represents the association between a campaign and a workflow
    properties:
      campaign:
        allOf:
        - $ref: '#/definitions/models.Campaign'
        description: Campaign represents the campaign associated with this workflow
      campaign_id:
        description: CampaignID is the ID of the campaign (foreign key)
        type: integer
      campaign_workflow:
        allOf:
        - $ref: '#/definitions/models.CampaignWorkflow'
        description: CampaignWorkflow represents the workflow associated with this
          campaign
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      workflow_id:
        description: WorkflowID is the ID of the workflow (foreign key)
        type: integer
    type: object
  models.CampaignWorkflowUser:
    description: Represents a user associated with a workflow and their status
    properties:
      campaign_workflow:
        allOf:
        - $ref: '#/definitions/models.CampaignWorkflow'
        description: CampaignWorkflow represents the workflow associated with this
          user
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      status:
        description: Status indicates whether the user is active or inactive in the
          workflow
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: User represents the user associated with this workflow
      user_id:
        description: UserID is the ID of the user (foreign key)
        type: integer
      workflow_id:
        description: WorkflowID is the ID of the workflow (foreign key)
        type: integer
    type: object
...
  models.Permission:
    description: A permission assigned to roles
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        description: |-
          Name is the unique name of the permission
          @Property name string true "Permission Name" example("view_users")
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    description: Role assigned to users with permissions
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        description: |-
          Name is the unique name of the role
          @Property name string true "Role Name" example("admin")
        type: string
      permissions:
        description: |-
          Permissions represents the permissions assigned to the role
          @Property permissions array "Permissions" items={Permission} example([{"id":1,"name":"view_users"}])
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
  models.User:
    description: Represents a user in the system with roles and campaigns
    properties:
      campaigns:
        description: Campaigns represents the campaigns associated with the user
        items:
          $ref: '#/definitions/models.Campaign'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        description: Email is the user's email address
        type: string
      id:
        type: string
      roles:
        description: Roles represents the roles assigned to the user
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      username:
        description: Username is a unique identifier for the user
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for My Project.
  title: My Project API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with username and password to get a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with username, email, and password
      parameters:
      - description: User registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the created user
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /campaign-recipient:
    post:
      consumes:
      - application/json
      description: Create a new campaign recipient with campaign ID, recipient ID,
        and status
      parameters:
      - description: Campaign recipient details
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CampaignRecipient'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new campaign recipient
      tags:
      - Campaign Recipients
  /campaign-recipient/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a campaign recipient by ID
      parameters:
      - description: Campaign Recipient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete an existing campaign recipient
      tags:
      - Campaign Recipients
    put:
      consumes:
      - application/json
      description: 'Update a campaign recipient by ID with optional fields: campaign
        ID, recipient ID, and status'
      parameters:
      - description: Campaign Recipient ID
        in: path
        name: id
        required: true
        type: string
      - description: Campaign recipient details
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CampaignRecipient'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update an existing campaign recipient
      tags:
      - Campaign Recipients
  /campaign-recipients:
    get:
      consumes:
      - application/json
      description: Get all campaign recipients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CampaignRecipient'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all campaign recipients
      tags:
      - Campaign Recipients
  /campaign-recipients/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a campaign recipient by ID
      parameters:
      - description: Campaign Recipient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CampaignRecipient'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get campaign recipient details
      tags:
      - Campaign Recipients
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"